gl basic method inheritence preservation.
I have created for loops that will hopefully
preserve the class type of the object that went into
the "[" method.

The "[" method for genlight and SNPbin object would
create new objects every time they were subset. This
prevents these objects from being properly inherited
because everything returned from that method will be
an object of class "genlight" or "SNPbin". This breaks
inheritence because any new class created from
genlight or SNPbin objects will not be able to use
callNextMethod() for the "[" method because the class
of object returned will no longer be the one that went
in.