Update handling.R
When working with simulated data it is possible that all our data is homozygous, in such cases when we asked for a summary of this data (as part of some data analysis pipeline) we could run into errors, namely "Error in out[, colnames(x@tab), drop = FALSE] : subscript out of bounds". This specific error can be fixed if in the tab function we specify that we never want to simplify the tapply output.